Moved ChainedServerInfo into its own package for easier reuse in config-server
